+++ src/main/java/de/is24/deadcode4j/plugin/packaginghandler/WarPackagingHandler.java
@@ -11,9 +11,10 @@ import org.codehaus.plexus.util.xml.Xpp3Dom;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 import java.io.File;
+import java.util.List;
 
+import static com.google.common.collect.Lists.newArrayList;
 import static de.is24.deadcode4j.Utils.getKeyFor;
-import static java.util.Collections.singleton;
 import static org.apache.commons.io.filefilter.FileFilterUtils.*;
 
 /**
@@ -27,20 +28,16 @@ public class WarPackagingHandler extends PackagingHandler {
     @Nullable
     @Override
     public Repository getOutputRepositoryFor(@Nonnull MavenProject project) throws MojoExecutionException {
-        final File webappDirectory = calculateWebAppDirectory(project, true);
-        return new Repository(new File(webappDirectory, "WEB-INF/classes"));
+        final File webAppDirectory = calculateWebAppDirectory(project, true);
+        return new Repository(new File(webAppDirectory, "WEB-INF/classes"));
     }
 
     @Override
     @Nonnull
     public Iterable<Repository> getAdditionalRepositoriesFor(@Nonnull MavenProject project) throws MojoExecutionException {
-        final File webappDirectory = calculateWebAppDirectory(project, false);
-        final File directory = new File(webappDirectory, "WEB-INF");
-        IOFileFilter fileFilter = notFileFilter(or(
-                        asFileFilter(new SubDirectoryFilter(directory, "lib")),
-                        asFileFilter(new SubDirectoryFilter(directory, "classes")))
-        );
-        return singleton(new Repository(directory, fileFilter));
+        List<Repository> repositories = newArrayList(getJavaFilesOfCompileSourceRootsAsRepositories(project));
+        repositories.add(getWebInfDirectory(project));
+        return repositories;
     }
 
     private File calculateWebAppDirectory(MavenProject project, boolean log) throws MojoExecutionException {
@@ -67,6 +64,17 @@ public class WarPackagingHandler extends PackagingHandler {
                     " does not exist - please make sure the project is packaged!");
         }
         return webappDirectory;
+    }
+
+    private Repository getWebInfDirectory(@Nonnull MavenProject project) throws MojoExecutionException {
+        final File webAppDirectory = calculateWebAppDirectory(project, false);
+        final File directory = new File(webAppDirectory, "WEB-INF");
+        IOFileFilter fileFilter = notFileFilter(or(
+                        asFileFilter(new SubDirectoryFilter(directory, "lib")),
+                        asFileFilter(new SubDirectoryFilter(directory, "classes")))
+        );
+        return new Repository(directory, fileFilter);
+
     }
 
 }