+++ src/test/java/de/is24/deadcode4j/analyzer/typeerasure/PackageClass.java
@@ -1,0 +1,3 @@
+package de.is24.deadcode4j.analyzer.typeerasure;
+@SuppressWarnings("UnusedDeclaration")
+public class PackageClass { }